requirements.txt
langextract
python-dotenv
pydantic